Fix compilation without gnutls
* imap4d/imap4d.c [WITH_TLS] (cb_tls_required): Conditional compilation.
* imap4d/imap4d.h [!WITH_TLS] (tls_encryption_on): Empty define.
* pop3d/pop3d.c [!WITH_TLS] (tls_available): Empty define.
* libmu_scm/mu_mailbox.c (mu_scm_mailbox_print): Fix format string.

+++ libmu_scm/mu_mailbox.c
@@ -81,7 +81,7 @@ mu_scm_mailbox_print (SCM mailbox_smob, SCM port, scm_print_state * pstate)
 	  
 	  scm_puts (p, port);
 	  
-	  snprintf (buf, sizeof (buf), " (%d)", count);
+	  snprintf (buf, sizeof (buf), " (%lu)", (unsigned long) count);
 	  scm_puts (buf, port);
 	}
       else
